+++ src/routes/+layout.svelte
@@ -70,6 +70,7 @@
 		apiKey={$settings.api_key}
 		maxConcurrentRequests={$settings.max_concurrent_requests}
 		checkInBackground={$settings.check_in_background}
+		noWindowDecorations={$settings.no_window_decorations ?? false}
 		bind:visible={$settingsOpen}
 	/>