+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Use Shift instead of Alt for Videos/Channels page shortcuts
+- Add option to disable window decorations
 - Fix keyboard input working when modal is open
 - Remove old YouTube Email Notifier migration